Rename `auto.new`
